Add stimulus.py to stimulus module and make grating stimulus function
Relying on data_objects and units, this function is now much simpler
than it used to be as most of the work previously was in converting
units and dealing with the orientation.

The direction of modulation is now orientation - 90 degrees, as is my
personal convention.

Converting from orientation to direction is done in the data_object
for grating parameters.

To ensure that the "wave" of the grating moves "forward" in the temporal
dimension, which is to say that the "wave" moves toward positive infty
in the spatial plane, the temporal coords are subtracted within the cos
function.

This works as it maps any given spatial_coordinate to a lower one.
IE, 12 degrees, after 3 seconds, becomes 9 degrees.
To then get the same point of the wave requires greater spatial coords.
IE, 15 degrees, after 3 seconds, will have the same part of the wave as
12 seconds did earlier.