fixed (?) int float issue with generics by just setting val_gen to float
Makes sense, should have had that instead of int and float given that
typevars are invariant.  In order to comply with the float>int
"shortcut" of the typing system (PEP 484), specifying only float is prob
necessary.

Other main motivation here was that setting the typevar to be covariant,
while it fixed this int/float issue, caused other issues, including that
function parameters couldn't be typed as val_gen, as covariant typevars
are not permitted (by pylance/pyright, not sure about mypy).