Add orientation biases parameters to DOG spatfiltargs and fitting funcs
Uses a specific function based on a definition of how to measure
circular variance from a cell.  In this case, Naito (2013).

This circular variance function is used to determine the circular var
of the base DOG spatial filter when stretched according to a ratio of
major to minor axes.

The ratio that produces a circular variance of ~0.99 is found through
automated minimisation.  Ratios from this max ratio to 1 (linspace)
are used to calculate corresponding circular variance values.  These two
arrays are then used to instantiate a new ori_bias_params attribute
of the filter (which is itself defined by a data object).

Methods of this object allow, through interpolation, calculation of
circ_var or ratio values from either.