Correct DC estimation
Previously it was multiplying the fourier values by the amplitude
of the stimulus.  **It should be the DC of the stimulus**.

This was a mistake from my prototyping that didn't caught because
the stimuli used in testing and prototyping were always had a
amplitude and DC of `1` so that the whole stimulus ranged from `0`-`2`.