Add Ground work for creating synchrony amongst LGN Cells
Calculate overlapping regions and use relative volumes of overlap
as weights on the temporal responses of cells.

Then add code to the management of multiple trials of multiple LGNlayers
to allow for a variable amount of cells in any given layer, as the
number of overlapping regions random and so varies.